feat: final submission of auction system

+++ database/migrations/2025_10_30_080237_create_images_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('product_id')->constrained()->cascadeOnDelete();
             $table->string('filename');
             $table->boolean('is_primary')->default(false);
-            $table->unsignedInteger('sort_order')->default(0);
+            $table->smallInteger('sort_order')->default(0);
             $table->timestamps();
             $table->index(['product_id','is_primary']);
             $table->index(['product_id','sort_order']);
